move `title_bar_style_ to `NativeWindow`
